requirements.txt
tensorflow==2.12.0
opencv-python
matplotlib
pandas
ipykernel
numpy
gdown
dvc
mlflow==2.2.2
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
Flask
Flask-Cors